refactor(dashboard): rename for clarity — /waterfall route, asset base, selector files (4/4)
Now that 'Dashboard' means the whole app, the leftover '/dashboard' naming
was confusing. Rename for clarity:
- route /__scenetest/dashboard -> /__scenetest/waterfall (it's the Waterfall
  view, not 'the dashboard').
- asset base /__scenetest/dashboard/ -> /__scenetest/ (assets at
  /__scenetest/assets/); middleware serves the assets prefix + the view paths.
- store.ts -> select-waterfall.ts, runner-store.ts -> select-runner.ts (they're
  selectors over the collection rows, not 'stores'); store.test.ts ->
  select-waterfall.test.ts. Imports + the e2e/middleware tests updated.
- CLAUDE.md Dashboard section synced to the modernized state (TSX, useLiveQuery,
  <a> nav, /waterfall, selector names).

Verified: dashboard + vite-plugin build/typecheck, 809 unit tests, and a dev
server serves /__scenetest, /runner, /waterfall + /__scenetest/assets/.
