+++ CLAUDE.md
@@ -79,11 +79,14 @@ The receiver core extracted from the Vite middleware, framework-agnostic so scen
 
 **Nomenclature:** the **Dashboard** is the *entire* Preact app and all its views — **Home**, **Runner**, **Waterfall**. (There is no "Console" — that name is retired.) Every view reads from one read-only `@tanstack/db` read model (the collections); there is no bespoke fold anymore.
 
-- `dashboard.ts` — `mountDashboard(element, { transport, theme?, base? })`: the whole app. Owns a shadow root + namespaced `dashboard-styles.ts`, routes on `location.pathname` (`/__scenetest` Home · `/__scenetest/runner` Runner · `/__scenetest/dashboard` Waterfall). **`useDashboardRows(transport)` builds the four collections** (`scenesProjection`/`assertionsProjection`/`actionsProjection`/`runsProjection` via `createCollection` + `runCollectionOptions`, sharing one `createRunSource`) and reads them reactively (`subscribeChanges` → re-render); every view gets the rows. The Waterfall view (`WaterfallHost`) renders into a **nested** shadow root (style-isolated from the chrome) but its data is the same store. Dev + cloud mount this same app; only the transport differs. Replaced `vite-plugin/src/analyze-app.ts` (deleted).
-- `runner.ts` — the Runner view (htm): `RunnerView` + `Tree`/`ListPane`/`Detail`/`SpecSnippet`/`CopyButton`. Live = `selectSnapshot(rows)`; past runs map a `/__scenetest/runs/:id` report.
-- `runner-store.ts` — Runner selectors: `selectSnapshot(scenes, assertions, actions, runs)` derives per-scene assertions/timeline from the collection rows via `attributeToScene`; `mapReportToSnapshot()` adapts a past-run JSON report into the same shape.
-- `app.ts` — `Waterfall` view component (htm) `{ state, send }` + `Header`/`SceneCard` + `sceneSummary()`. A pure view — the root passes `state` (from `selectWaterfall`).
-- `store.ts` — `selectWaterfall(scenes, assertions, actions, runs)`: derives the Waterfall's `DashboardState` (lanes by actor, attributed assertions, run rollup) from the collection rows. (Replaced the old `applyEvent` event-fold.) + `completedSceneCount`, `initialState`.
+The views are **TSX/JSX** (the package has a build step, so no htm), and read the store reactively with **`useLiveQuery`** (`use-live-query.ts`) — no hand-rolled `subscribeChanges`. Navigation is real `<a>` links, intercepted *inside* the shadow root (no preact-iso: its document-level delegation misses shadow-retargeted `<a>` clicks).
+
+- `dashboard.tsx` — `mountDashboard(element, { transport, theme?, base? })`: the whole app. Owns a shadow root + namespaced `dashboard-styles.ts`, routes on `location.pathname` (`/__scenetest` Home · `/__scenetest/runner` Runner · `/__scenetest/waterfall` Waterfall). **`useDashboardStore(transport)` builds the four collections once** (`scenesProjection`/`assertionsProjection`/`actionsProjection`/`runsProjection` via `createCollection` + `runCollectionOptions`, sharing one `createRunSource`) and passes them down; the root only tracks connection liveness. Each view reads the tables it needs via `useLiveQuery`, so re-renders are scoped. The Waterfall view (`WaterfallHost`) renders into a **nested** shadow root (style-isolated from the chrome) but its data is the same store. Dev + cloud mount this same app; only the transport differs. Replaced `vite-plugin/src/analyze-app.ts` (deleted).
+- `runner.tsx` — the Runner view: `RunnerView` + `Tree`/`ListPane`/`Detail`/`SpecSnippet`/`CopyButton`. Live reads the collections via `useLiveQuery` → `selectSnapshot`; past runs map a `/__scenetest/runs/:id` report. (Text filter + section grouping stay reactive selectors over the live rows — groupBy yields aggregates, not bucketed lists.)
+- `select-runner.ts` — Runner selectors: `selectSnapshot(scenes, assertions, actions, runs)` derives per-scene assertions/timeline from the collection rows via `attributeToScene`; `mapReportToSnapshot()` adapts a past-run JSON report into the same shape.
+- `app.tsx` — `Waterfall` view component `{ state, send }` + `Header`/`SceneCard` + `sceneSummary()`. A pure view — `WaterfallHost` passes `state` (from `selectWaterfall`).
+- `select-waterfall.ts` — `selectWaterfall(scenes, assertions, actions, runs)`: derives the Waterfall's `DashboardState` (lanes by actor, attributed assertions, run rollup) from the collection rows. (Replaced the old `applyEvent` event-fold.) + `completedSceneCount`, `initialState`.
+- `use-live-query.ts` — `useLiveQuery(collection)`: subscribe a TanStack DB collection (base or `createLiveQueryCollection`) and read its rows reactively. The Preact analogue of `@tanstack/react-db`'s hook.
 - `select-helpers.ts` — `latestRunSlice(...)` (the multi-run store sliced to `where runId = latest`, shared by both selectors) + the `DashboardRows` type.
 - `dev-transport.ts` — `createDevTransport()`: fetch + SSE adapter; maps `Command`s to `/__scenetest/replay|stop|pause`
 - `types.ts` — `Transport`, `DashboardState`, `Scene`, `DashboardTheme`, `ConnectionStatus`, `MountOptions`
