+++ packages/dashboard/src/index.ts
@@ -1,7 +1,7 @@
 export { mountDashboard } from './dashboard.js'
 export { createDevTransport, type DevTransportOptions } from './dev-transport.js'
-export { selectWaterfall, initialState, completedSceneCount } from './store.js'
-export { selectSnapshot, mapReportToSnapshot } from './runner-store.js'
+export { selectWaterfall, initialState, completedSceneCount } from './select-waterfall.js'
+export { selectSnapshot, mapReportToSnapshot } from './select-runner.js'
 export { sceneSummary } from './app.js'
 export { useLiveQuery, type ObservableRows } from './use-live-query.js'
 export type { DashboardCollections } from './select-helpers.js'
